% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgis-package.R
\docType{package}
\name{arcgis-package}
\alias{arcgis}
\alias{arcgis-package}
\title{arcgis: ArcGIS Location Services Meta-Package}
\description{
Provides easy installation and loading of core ArcGIS location services packages 'arcgislayers', 'arcgisutils', 'arcgisgeocode', and 'arcgisplaces'. Enabling developers to interact with spatial data and services from 'ArcGIS Online', 'ArcGIS Enterprise', and 'ArcGIS Platform'. Learn more about the 'arcgis' meta-package at \url{https://r.esri.com/r-bridge-site/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R-ArcGIS/arcgis/}
  \item \url{https://r.esri.com/arcgis/}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
