% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{arcgis_packages}
\alias{arcgis_packages}
\title{List all {arcgis} packages}
\usage{
arcgis_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{default \code{TRUE}. Includes the arcgis package name in the
resultant character vector()}
}
\value{
A character vector of package names included in the arcgis metapackage.
}
\description{
List all {arcgis} packages
}
\examples{
if (interactive()) {
  arcgis_packages()
}

}
